sn: add context and internal timeout to LogStreamReporter (#134)
This commit adds timeouts to LogStreamReporter's GetReport and Commit.
To do that, it uses context package. They are using channels to enqueue
GetReport and Commit requests. Because channels in golang are bounded,
there are chances to pause handling requests, even very long times.

Before this change, LogStreamReporter's GetReport and Commit have no
preemption points, that is, once they are called, there are no ways to
cancel the operations. This change gives preemption points by using the
context package.

Fixes #132