sn,lse: track appendT before incrementing uncommittedLLSNEnd (#188)
This patch fixes very subtle bug; client appending a log can be stuck.
Before this patch, LSE.write method increments uncommittedLLSNEnd before
tracking the appendTask while processing an append request. In this
case, however, LSE can report a new uncommitted log, which is not
tracked yet. It is also possible to receive commit the not-yet-tracked
log. Committing the log is okay, but it misses notifying the commit to
the client.

This change has simple fix. LSE.write tracks the appendTask after
succeeding writing the log to the storage, and then, increments
uncommittedLLSNEnd. Tracking the appendTask just adds the task to the
map for appendTasks. Since not-yet-uncommittedLLSNEnd-incremented task
won't be reported, it won't be harmful.

Fixes #VARLOG-223