mcl: implement Seal API (#199)
* mcl: implement AddLogStream API

This implements AddLogStream API in management server. It supports two
types of API:

- add logstream manually: specify storage nodes and logstream id to add
- add logstream simply: let vms choose storage nodes and logstream id

In some case, there can be a logstream id added to storagenode but not
registered to MR. This situation can be happened by partially processed
AddLogStream. In this case, AddLogStream returns
ErrLogStreamAlreadyExists, and it is transient error.

Fixes #VARLOG-232

* mcl: implement Seal API

This patch implements Seal API in management server. First, it calls
Seal API of MR, and then it calls Seal API of every SNs those have the
given logstream.
This includes quite a simple seal test.

Fixes #VARLOG-233

* fix deadlock issue: SetDirty won't acquire mutex

* MRManager implements ClusterMetadataView for clear and simple
implementation.