mcl: implement RemoveLogStreamReplica API (#216)
RemoveLogStreamReplica removes a log stream replica in a given storage
node. To remove the log stream replica, it should be satisfied with the
follows; log stream replica to delete should not be contained to log
stream manged by metadata repository. There are several situations, for
example:

- The log stream replica maybe garbage which is created but not
registered to MR.
- The log stream to which the log stream replica belongs might be
unregistered.
- The log stream replica does not belong to the log stream which is
registered to MR since UpdateLogStream API excluded it.

It calls RemoveLogStream RPC to the storage node internally. It results
in a closing log stream executor in the storage node. Optionally, it
removes the directory corresponded to the log stream, but it is not
implemented yet.

Resolves #VARLOG-255