mcl: change statRepository's state when called UpdateLogStream (#228)
* mcl: change statRepository's state when called UpdateLogStream

Assumes that all the replicas in a log stream are LogStreamStatusSealed.
The status of the log stream of StatRepository is also
LogStreamStatusSealed. To exchange a replica in the replicas,
UpdateLogStream is called. Successful UpdateLogStream should make the
status of the log stream non-LogStreamStatusSealed. To do that, this
patch changes the status of the log stream in StatRepository after
successfully calling UpdateLogStream.

This PR has some unrelated fixes, for example, acquiring mutex for
public API in the cluster manager. To avoid deadlock when closing VMS,
some functions are added contexts as parameters.

Resolves #VARLOG-266

* fix: code review

* resolve deadlock when closing vms