cl: start implementing varlog client (#235)
* *: move pkg/varlog/types to pkg/types

To be clear the purposes of the packages:

- The package `types` is a base package that is used for other packages.
- The package `varlog` is to provide client library to end-users.

Other packages which need only `types` need not to depend on `varlog`.

* *: move pkg/varlog/util to pkg/util

* cl: remove unused files

* cl,sn: move LogEntry to pkg/types

* *: move rpc_client.go to pkg/rpc

* *: move errors.go to pkg/verrors

To avoid possible import-cycle, errors.go is moved from pkg/varlog to
pkg/verrors. Data types and constants in errors.go are used widely, thus
it can be source of import-cycle easily.

* mrc: move clients for mr to pkg/mrc

The package `pkg/mrc` contains clients for the metadata repository. The
users of `pkg/varlog` can connect to the metadata repository via the
connector of `pkg/mrc` rather than directly. The connector will be
implemented further patches.
Moving clients for the metadata repository to `pkg/mrc` also can avoid
possible import-cycle by preventing unnecessary import `pkg/varlog`.

The `mrc` is abbreviate for `metadata repository clients`.

* snc: move clients for storage node to `pkg/snc`

The package `pkg/snc` contains clients for the storage node. The reason
why this change is similar to the reason for moving clients for the
metadata repository to `pkg/mrc`.

* mrc: add mrconnector to use in both mr_manager and varlog client

The package `pkg/mrc/mrconnector` is a common code for both mr_manager
in vms and varlog client. In this patch, MRManager is refactored by
using mrconnector.

* cl: start implementing varlog client

This patch is starting point of implementing the varlog client.

Resolves #VARLOG-271

* cl: fix code review

- the order of update rpc address in mrconnector
- add missing ticker to expireDenyTTL in allowlist
- apply singleflight to warmup in allowlist
- do rafactoring log_io_client_manager (separate package & apply
  singleflight)