cl: implement Subscribe API (#259)
This patch implements Subscribe API in varlog client. It just receives a
range of logs. In the next steps, it will be extended to prefetch.

Subscribe spawns several goroutines, for instance, `subscriber`s, a
`transmitter`, a `dispatcher`. `Subscriber`s call RPCs to LogIOService
to receive logs and put the logs into the internal queue. `Transmitter`
checks internal queues of `subscriber`s and moves the proper log from
the `subscriber` to `subscribedLogEntriesQueue`. `Dispatcher` calls
user-defined callback function repeatedly.

To lower CPU usage of `Transmitter`, `Subscriber`s notify new incoming
logs to `Transmitter` by using a channel (e.g., transmitCV).
`Dispatcher` calls user-defined callback function with `io.EOF` to
announce the end of subscribed logs. The client can stop by using closer
which is returned by Subscribe API, and the callback function may get an
error as a parameter when the log stream is stopped in the middle.

This includes unit tests and integration tests for subscribe.

Resolves #VARLOG-281