sn: fix log stream report's bug (#281)
* sn: fix log stream report's bug

This patch fixes the bug of log stream report and changes algorithm of
report:

- If new LSEs exist, thus their HWMs are zero, the report contains only
them. MR is not likely to have very old GLS of others.
- The history of the report is not maintained by the level of the
storage node. Rather, it is maintained by the level of log stream
executor.
- When the HWM of LSE's status is not the same as the HWM of the report,
it doesn't calibrate the uncommitted length of LSE's status. Instead,
when the HWM of LSE's status is equal to the element in history and the
uncommitted length of it becomes longer, the new status of LSE is
reported.

Especially, the last is important. Before applying this patch, when the
LSE did commit just ago, the number of uncommitted log entries (a.k.a,
the length of uncommitted) of the report can be the sum of the number of
uncommitted log entries of the previous status and recently written
logs. At that time, the HWM of the status in the report is equal to the
previous status.

```
// Before this patch

<previous status of LSE-10>
lsid=10
hwm=380
uncommitted_llsn_offset=41
uncommitted_llsn_end=47

<actual status in the report>
REPORT's hwm=380
lsid=10
hwm=380
uncommitted_llsn_offset=41
uncommitted_llsn_end=47

<after commit and writing 2 more log entries, status of LSE-10>
lsid=10
hwm=390
uncommitted_llsn_offset=44
uncommitted_llsn_end=49

<actual status in the report>
REPORT's hwm=380
lsid=10
hwm=380
uncommitted_llsn_offset=41
uncommitted_llsn_length=49
```

The last report of `LSE-10` is calibrated to the previous phase and its
uncommitted length is enlarged to 9. But it is meaningless since it has
already GLS in MR whose commit range is `[41,43]`.

After this patch, the last report of `LSE-10` is equal to the previous
one (e.g., `[41-47]`). If it is not committed and its uncommitted length
is increased, the new report (e.g., `[41-49]`) is used.

If you want reactivate report calibration, see `log_stream_reporter.go`.
Simply removing the comments can reactivate the report calibration.

Resolves #VARLOG-322

* changes:

- use report calibration
- includes status which have invalid hwm into the report