log: use tracer (#292)
This patch adds tracer to varlog. The tracer uses opentelemetry, for
example, this patch uses jaeger to visualize the traces.
It also includes a new logger to use tracer. To use this new logger, the
whole codebase should be changed, so this patch just adds it, but not
uses it.

Resolves #VARLOG-330