mr,mrc: mr_connector gets the addresses of only live MRs (#305)
Before this patch, mr_connector managed all of the addresses of MRs
including failed nodes. It, however, resulted in unnecessary connection
errors.  This patch checks if the MR is actually connectable by calling
GetMetadata RPC. It also includes options to set
`ClusterInfoFetchInterval` of the connector.

Note that this patch doesn't consider whether the connected MR is
partitioned or not. It must be a very important issue, but I'm not sure
that the mr_connector should care about this issue.

Resolves #VARLOG-341