mrc: mr connector use seeds and last connected node if needs (#338)
Sometimes, mr connector can lose all of mr connections. This patch uses
the seeds and last connected node whenever the mr connector loses all of
the connections to mr.
MR Connector also checks if the connected node is partitioned. To
confirm that, it uses `GetClusterInfo` RPC call, and check that
`ClusterInfo.Leader` is set. MR can achieve it that the MR unsets
`ClusterInfo.Leader` field when it is partitioned.

Resolves #VARLOG-377