sn: restore context after restarting executor (#364)
This patch enhances recovery after restarting the log stream executor.

Before this patch, a log stream executor assumed that it knows all of
the committed logs before. It, however, was wrong, since the executor
never can know until what position the logs were committed before the
executor crashed. For instance, the last commit message sent from MR
could be lost due to the failure of the executor.

In this change, the executor recovers its log entries by using local
storage and the seal RPC. At first, the executor starts with a SEALING
state. To append a log to the executor, it should be unsealed and change
its status to RUNNING.  To do so, two RPCs are needed - seal and unseal.
Seal RPC tells the executor which is the last committed log position.
From this RPC, the executor can recover all committed logs.

You can see the detail of the design from this URL:
https://docs.google.com/presentation/d/1FLvjqjrkfNHvZcxBLmLCQPsas7P-C6Z7NWGqLv8ukDQ/edit?usp=sharing

Resolves #VARLOG-412