mr: Even if sealed LS exists, CommitResultHistory should be trimmed (#409)
This patch allows trim on CommitResultHistory even when some LS is sealed.

MR manages LSs as `Sealing` state that have not yet reached sealedGLSN 
among LSs that have received a Seal request. When there is a condition to 
change the LS status to `Sealed`, the MR seals the `ReportCollector` to prevent 
the commit to the SN. That allows MR trims `CommitResultHistory`.

When unseal LS, MR unseals the `ReportCollector` with current highWatermark.
`ReportCollector` sets `commitBeginHighWatermark` to the highWatermark and
use it as a begin highWatermark for committing. This ensures that the already 
trimmed `CommitResultHistory` is not accessed.

See https://wiki.daumkakao.com/pages/viewpage.action?pageId=772862473

Resolves #VARLOG-464