sn: ignore Commit RPC in SEALED state (#407)
This patch changes the behavior of the Commit RPC handler of the log
stream executor when it is SEALED state. Before this change, the log
stream executor processed the Commit RPC when it is SEALED. Concretely,
it expected that the given Commit RPC must be empty when SEALED since it
had to know the global HighWatermark even if there was no log entry to
commit. Now, it is not necessary to know the global HighWatermark when
the log stream executor is SEALED.

See: https://wiki.daumkakao.com/pages/viewpage.action?pageId=772862473#id-05Report&Commit-Approach-002

Resolves #VARLOG-472