sn: block commit when either learning or sealed (#415)
This patch prevents log stream from committing logs when it is either
state of LEARNING or SEALED. It is also enhancing method for drainage of
commitWaitTasks called when is about to be SEALED. In addition, this
patch implements a new method - waitForDrainageOfCommitTasks. It makes
commitTaskQ be empty. It is necessary when the log stream is about to do
sync operation.

Resolves #VARLOG-474