sn: fix response condition of backup replication (#420)
Request of Backup replication can be replied after writing the logs to
the storage. Before this patch, it replied after committing the logs in
backup replica. It was not necessary to wait for committing the
replicated logs in backup replica to reply the replication request. This
patch fix this, thus now the backup replica replies immediately response
of backup replication after writing the logs into the storage.

This patch contains the above fix as well as some other renames -
refactoring. Struct appendTask is renamed to writeTask. It is more
proper name. Committer use commitWaitTask rather than appendTask. Struct
commitWaitTask is a new type to represent element of commitWaitQ.

Resolves #VARLOG-467