sn: use value of LogStreamUncommitReport in reports (#444)
This patch changes type of field `UncommitReports` in
`proto/snpb/GetReportResponse` from the slice of pointer of
`proto/snpb/LogStreamUncommitReport` to slice of the value of it. Since
StorageNodes create a very huge amount of reports, allocations of lots
of heap objects are very burdensome. This change, thus, is expected to
lower GC-overhead.

Resolves [#VARLOG-511](https://jira.daumkakao.com/browse/VARLOG-511).