sn: not making map/slice to get reportcommitters (#451)
When a StorageNode has many LogStreams, it costs a lot to get a list of
ReportCommitters, since getting a list of ReportCommitters makes a new
map and slice. To handle GetReport RPC, StorageNode calls a function to
get a list of ReportCommitters. Thus it takes quite CPU time.

This patch adds a new method to traverse a list of ReportCommitters
without making any slice and map.

The `internal/storagenode/reportcommitter.Getter` interface now has two
new methods:

- `ForEachReportCommitter(func(ReportCommitter))`
- `NumberOfReportCommitters() int`

Previously, to lookup all ReportCommitters or to get the number of
ReportCommitters, it makes a new map and slice. However, the above
methods won't.

Resolves [#VARLOG-514](VARLOG-514).