sn: single LogStreamCommitResult in CommitRequest (#473)
* sn: single LogStreamCommitResult in CommitRequest

This patch changes the number of LogStreamCommitResult in CommitRequest;
from multiple results to a single result. It also changes the type of
the field: from slice to a value.

The size of `proto/snpb/LogStreamCommitResult` is 48 Bytes. I am not
sure whether the value type is more efficient than the pointer type. I
will experiment with that.

Resolves [#VARLOG-532](VARLOG-532).

* change methods of `internal/storagenode/reportcommitter.(Client)`

- GetReport does not have context parameter.
- Commit does not have context parameter.
- Commit receives value type of commit request.