sn: add memtable-related options to storage (#448)
This patch adds `MemTableSize` and `MemTableStopWritesThreshold` options
to the `internal/storagenode/storage`. They are private options, that
is, they are configured only for experiments and internal routines. The
default value of `MemTableSize` is 4M, however, in storagenode
executables, its value is set to 64M.
It also contains new configurations `--storage-debug-log`, which prints
out events of storage.

Resolves [#VARLOG-513](https://jira.daumkakao.com/browse/VARLOG-513).