sn: set reason of seal (#501)
* sn: set reason of seal

Before this change, there is no way to know the reason why a log stream
replica is sealed. This patch sets the reason why the status of the log
stream replica sets to SEALING.

When the log stream is starting, its status is SEALING. If the client
requests append RPC, then the log stream replica returns an error -
ErrSealed with reason stating "initial state".

When the log stream received seal RPC when it is running (aka mutable),
its status becomes "SEALING" or "SEALED" according to its internal
status. If the client requests append RPC again, then the log stream
replica returns an error - ErrSealed with reason stating "seal rpc".

The reason for sealing should be used for debugging and monitoring. We
must not use its text to decide or branch execution flows.

Resolves [#VARLOG-578](https://jira.daumkakao.com/browse/VARLOG-578).

* add a log to debug TestVarlogFailoverSNBackupFail