test: simple varlog for testing (#513)
* admin,client: refactoring `pkg/varlog`

This patch does refactoring of a client of admin. It now moves to
package `pkg/varlog`. The return types of its methods are changed not to
use the response of RPCs directly since it is very inconvenient for
users.

Refactoring of admin client is still work-in-progress. There are still
many changes necessary, for instance, more precise method naming, more
practical RPC signatures as well as many features of an admin client
application.

It renames `pkg/varlog.Varlog` to `pkg/varlog.Log` because the former is
more awkward.

Resolves [#VARLOG-591](VARLOG-591).

* test: simple varlog for testing

The package `github.com/kakao/varlog/pkg/varlogtest` is a
testing tool to test the varlog's operations.

Resolves [#VARLOG-570](VARLOG-570).

* remove unnecessary code