admin,client: refactoring `pkg/varlog` (#515)
This patch does refactoring of a client of admin. It now moves to
package `pkg/varlog`. The return types of its methods are changed not to
use the response of RPCs directly since it is very inconvenient for
users.

Refactoring of admin client is still work-in-progress. There are still
many changes necessary, for instance, more precise method naming, more
practical RPC signatures as well as many features of an admin client
application.

It renames `pkg/varlog.Varlog` to `pkg/varlog.Log` because the former is
more awkward.

Resolves [#VARLOG-591](https://jira.daumkakao.com/browse/VARLOG-591).