storagenode: fix data race when sealing (#602)
This patch fixes a data race issue that can occur when the log stream
replica is sealed. When the log stream replica is sealed, it should
drain internal queues, for instance, task queues of the sequencer, the
writer, and the committer, then reset their internal states. Throughout
these steps, the internal states are accessed concurrently, thus the
data race can occur.
This patch rearranges the fence to avoid this race.

Resolves [#VARLOG-691](https://jira.daumkakao.com/browse/VARLOG-691).