admin: refactoring varlogadm for tests (#666)
This patch contains the following changes:

- `internal/varlogadm.(StorageNodeManager)` is moved to
  `internal/varlogadm/snmanager.(StorageNodeManager)`.
- `internal/varlogadm.(MetadataRepositoryManager)` is moved to
  `internal/varlogadm/mrmanager.(MetadataRepositoryManager)`.
- `internal/varlogadm/snmanager.(StorageNodeManager)` and
  `internal/varlogadm/mrmanager.(MetadataRepositoryManager)` are
  injected to `internal/varlogadm.(ClusterManager)` rather than
  instantiating in the constructor of
  `internal/varlogadm.(ClusterManager)` to improve testability of varlog
  admin server.

`StorageNodeManager` and `MetadataRepositoryManager` were created in the
constructor of `ClusterManager`, which made it challenging to use
mockups for unit tests. Since they connect to other servers, for
instance, storage nodes and metadata repositories, we need to be mocking
them up to test easily. This patch, hence, changes how to instantiate
them.

Resolves [#VARLOG-658](VARLOG-658).