fix(admin): set empty slice to `logStreams` in GetStorageNode response rather than null when no log stream replicas
When there was no log stream replica in a storage node, the JSON field `logStreams` in the
GetStrorageNode response was null. It can cause an exception while accessing the field after
decoding the response, for example `TypeError: 'NoneType' object is not iterable`.

This patch fixed the above case by setting empty slice to the
`proto/snpb.(StorageNodeMetadataDescriptor).LogStreamReplicas`.

Resolves #88