fix(admin): contains all storage nodes registered to the cluster
RPCs GetStorageNode and ListStorageNodes should contain all storage nodes registered to the cluster
even if the admin server is just restarted.
Previously, if the admin server was restarted the stats repository of the admin was flushed, thus,
it had a problem that the RPCs do not return all storage nodes after restarting.
This patch fixes the problem by using the cluster metadata fetched from the metadata repository. If
the metadata repository is not reachable transiently, the RPCs return an error with the status code
`Unavailable`.

Resolves #106