fix(storagenode): fix a response of SyncInit when source replica was trimmed
This patch fixed a response of `/varlog.snpb.Management/SyncInit` to respect trimmed source replica.

The cause of the problem was that the destination replica replied` uncommittedLLSNBegin` of itself
as the starting position of synchronization regardless of the source replica's condition. When the
source replica has been trimmed, the first log position of the synchronization range can be larger
than the `uncommittedLLSNBegin` of the destination replica. In this case, the destination should set
its `uncommittedLLSNBegin` and `uncommittedLLSNEnd` to the first log position of the synchronization
range.
Note that this synchronization makes the destination replica's all logs be removed and filled with
the new logs from the source replica. To reflect that the source replica has no logs, the
`localLowWatermark` and `localHighWatermark` of it should be set to invalid by `SyncInit`. They will
be changed by a commit triggered by synchronization.

Resolves #134