fix(storagenode): fix error-prone state management of SyncInit and Report
When a log stream replica handles SyncInit, it can be prone to mutate the replica's state too early.
It should change the state after checking a valid sync range. It also updates the version, high
watermark, and uncommittedLLSNBegin with zeros to denote that the replica sends an invalid report
while the learning phase.
The replica checks whether it is in the learning phase at the end of method
`internal/storagenode/logstream.(*Executor).Report` to minimize the probability of concurrency
issues mentioned in #159.

Refs
- https://github.com/kakao/varlog/pull/159#issuecomment-1254917165