feat(storagenode): change the synchronization method to accept only the last commit context
This patch changes the synchronization process to accept only the last commit context applied by #162.
The synchronization copies log entries and the latest commit context from the source replica to the
destination replica. The destination's reportCommitBase should be valid, and its log entries should
be the same as the source's.

SyncInit can now delete log entries that the source replica had already removed. It deletes all log
entries simultaneously to avoid a hole in the middle of log entries. It also makes the replica's
reportCommitBase invalid because commit context and log entry are inconsistent during the
synchronization.

SyncReplicate no longer uses the SyncReplicateBuffer since there is no reason to store a commit
context and corresponding log entries in the buffer and write them into a disk at once. It now saves
log entries or the commit context every time. We can optimize the SyncReplicate to transfer a batch
of log entries, which we can discuss later.

Updates #125