perf(storage): use a lightweight method to get the data size of the storage
The previous method to get the data size of storage belonging to a log stream was costly. Concretely
it used too many heap objects to call `filepath.Walk` API. To avoid this problem, this patch changes
the implementation of `internal/storage.(*Storage).DiskUsage` to use the built-in method of pebble
DB - `DiskSpaceUsage`.

Resolves #210