feat(storagenode): add an upper limit of log stream replicas count in a storage node
This patch adds a new CLI flag, `--max-logstream-replicas-count,` which limits the number of log
stream replicas in a storage node. The default value is -1, meaning no upper limit, and a storage
node cannot create a log stream replica if it is zero.

A storage node keeps a count variable for the number of log stream replicas. Whenever the storage
node creates a new replica, it checks if the number of log stream replicas exceeds the upper limit.
If there is no problem, the storage node creates a new replica and increases the count variable.
Conversely, the storage node decreases the count variable whenever a log stream replica is removed.

This approach has a limitation: it cannot distinguish between standard and garbage replicas.
Fundamentally, a storage node has no method to distinguish them. So, this PR won't consider that.

Resolves #293