feat(metarepos): add an upper limit for the number of topics in a cluster
This patch adds an upper limit for the number of topics in a cluster. To do that, it adds a new CLI
flag `--max-topics-count` to the varlogmr executable.

When the metadata storage tries registering a new topic, it compares the current number of topics in
the cluster against the given upper limit. If the metadata repository rejects the RegisterTopic RPC,
gRPC code of the response is ResourceExhausted.

Resolves #295