feat(metarepos): add an upper limit for the number of log streams in a topic
This patch adds an upper limit for the number of log streams in a topic. The threshold can be set by
the `--max-logstreams-count-per-topic` flag to the metadata repository. If it is a negative value,
an unlimited number of log streams can be created per topic. Note that no log streams can be made if
it is zero.

When `internal/metarepos.(*MetadataStorage)` handles `registerLogStream`, it confirms whether the
number of log streams in the topic is greater than or equal to the upper limit.

Later, we can add a new feature, like a topic-specific upper limit of log stream count.

Resolves #297