feat(storagenode): add a new sync state - SyncStateStart
This patch adds a new sync state - `proto/snpb/SyncStateStart`. It indicates the start of a new
synchronization job.

Previously, there were only three states: SyncStateError, SyncStateInProgress, and
SyncStateComplete. If the admin server has called Sync RPC to a storage node, the node has replied
SyncStateInProgress if the synchronization started successfully. The admin server also could call
Sync RPC next time, and then the storage node replied SyncStateInProgress again. We had no method to
discriminate between the start and the progress of the synchronization job.

From the result of this patch, the response of Sync RPC creating a new synchronization job will have
SyncStateStart rather than SyncStateInProgress. It will be helpful for the admin server to manage
the life cycle of synchronization jobs.

Resolves #299