feat(admin): defines error codes of several RPCs in the admin server
This patch defines the gRPC error codes of various RPCs in the admin server. Previously, we depended
on the `verrors` package to identify the cause of errors. However, the `verrors` package is vague
and an anti-pattern. To handle errors explicitly, we are trying to adopt gRPC error codes.
Although this PR doesn't convert everything, it can be a starting point for our robust error
handling.

Updates #312