feat(storagenode): change SyncInit to handle trimmed source replica
This patch changes the behavior of SyncInit to handle empty source replicas. Previously SyncInit
thought that an empty source could not exist. Changes are the followings:

- Destination replica, SyncInit RPC server replies gRPC AlreadyExists error if it already has enough
  log entries and a commit context.
- Destination replica trims its local log entries if the source replica already did and receives a
  commit context if necessary.

Soon, we will change the behavior of Trim to be able to remove all log entries from a log stream
replica. This PR prepares it.

Updates #351