fix(storagenode): ignore context error while checking to interleave of Append RPC errors
Append batch requests can succeed partially. In case of partial failure, failed log entries must be
sequential. That is, suffixes of the batch can be failed to append, whose length can be from zero to
the full size.

Our codebase panics when the batch's success and failure are interleaved. This PR ignores context
errors caused by clients since the batch can succeed in the storage node, although clients canceled
it.