feat(storage): handle inconsistency between data and commit in ReadRecoveryPoints
Previously the method internal/(*storage).ReadRecoveryPoints returned an error when it found an
inconsistency between data and commit for log entries, which could happen when there was no data for
a commit. It should not have happened when we used the unified database in storage, so the method
returned an error.
However, using separate databases in storage and turning off the WAL sync option can happen; for
instance, the data for a log entry can be lost, although the commit for that log entry is synced.

This PR makes the ReadRecoveryPoints not return an error for the inconsistency between data and
commit for a log entry. It tries to find the first and last log entries that have no inconsistency.
If there are no valid first and last, it returns nil for them and lets them be resolved through
synchronization.