refactor(admin): prevent the large lock from mrmanager
This PR changes the role of `internal/admin/mrmanager.(*mrManager).mu`. The large lock in the
mrmanager made the mrmanager run sequentially. However, from now, it only guards the dirty flag for
the cluster metadata. Since the metadata repository works concurrently, adopting a large lock on the
client side is unnecessary.