fix(varlogtest): validate log stream descriptor to create it
This PR adds a validator to the AddLogStream API in the varlogtest.
Especially the validator checks the replicas requested. For example, if they have duplicate storage
nodes, their length is less than the replication factor, or they have unknown storage nodes.

Although the real varlogadm server's AddLogStream API is tested, the same thing in the varlogtest
package has not been tested, surprisingly. To fix it, all the implementations in the
`internal/admin` and `pkg/varlogtest` have to use the same test suite. However, it will lead to a
quite big refactoring. Therefore, this PR fixes the bug and adds some unit tests in the varlogtest
package.

Later we will revisit it to resolve #592.