perf(benchmark): use grpc.SharedBufferPool for grpc.DialOption
When a client library parses incoming gRPC messages, it allocates new buffers
frequently. This can result in the creation of a large number of heap objects.
However, gRPC offers an experimental feature called the shared buffer pool that
can be used to address this issue.

This pull request (PR) aims to optimize the gRPC receiving path using the shared
buffer pool. Various experiments have shown that this optimization leads to a
decrease in CPU time of runtime.gcBgMarkWorkers, depending on the workload.

Since the shared buffer pool is experimental, it is applied only to the
benchmark.

Refs:
- https://pkg.go.dev/google.golang.org/grpc@v1.62.0/experimental#WithRecvBufferPool
- https://pkg.go.dev/google.golang.org/grpc@v1.62.0#SharedBufferPool