feat(sn): add metrics for Append and Replicate RPCs
This pull request defines metrics for measuring RPCs such as Append and
Replicate. It introduces four metrics:

- `log_rpc.server.duration` measures the time spent processing inbound RPC calls
  in microseconds. It is very similar to the `rpc.server.duration` defined by
  OpenTelemetry, but our metric also measures the processing time triggered by
  each call on a gRPC stream.
- `log_rpc.server.log_entry.size` measures the size of appended log entries. It
  is similar to the `rpc.server.request.size` metric, but our metric measures
  the size of each log entry included in the appended batch.
- `log_rpc.server.batch.size` measures the size of log entry batches appended.
- `log_rpc.server.log_entries_per_batch` measures the number of log entries per
  appended batch.

These metrics are histogram-type, allowing us to compute percentiles and analyze
histograms and heat maps. Users can leverage these metrics to analyze the
duration of RPCs, the distribution of log entry sizes, and the length of
batches. We expect users to find better configurations to optimize storage node
performance.